moved classes to fewer files
