--- measurement/fastscan/__init__.py
@@ -1,29 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-
-@author: Steinn Ymir Agustsson
-
-    Copyright (C) 2018 Steinn Ymir Agustsson
-
-    This program is free software: you can redistribute it and/or modify
-    it under the terms of the GNU General Public License as published by
-    the Free Software Foundation, either version 3 of the License, or
-    any later version.
-
-    This program is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU General Public License for more details.
-
-    You should have received a copy of the GNU General Public License
-    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-"""
-
-
-def main():
-    pass
-
-
-if __name__ == '__main__':
-    main()
